Zynthian Encoder Display Power Fan (epdf)
Initial Commit for hw project:
This is a basic mcp23017 encoder, with 5 pin headers to align with
Arduino style encoders. Also includes power input 5V to drive the
Pi from a reliable power source. A display power connector for driving
a 7" Pi foundation screen. A PWM driven fan for Pi4 cooling. An
Activity LED and Reset SW to allow soft-boot from power down, and
an indication of a "busy" Zynth.