Move GPIOs change act_led R19 220 Ohm
Based upon various descriptions of Pi LED, and 3.3V Decided to use
220 Ohm resistor, rather than 0Ohm or 1K. 603 pad is possible to
replace resistor if required. Should be low enough R to work on all LED

Rework GPIO lines to use lines marked not in use by Ozz display in
hopes of keeping this compatible for most folks.

It will be possible to mag wire rework and epoxy glue to do a test
upgrade before the next rev of HW. Should even be clean and tidy.